+++ pybamm/solvers/casadi_solver.py
@@ -68,7 +68,7 @@ class CasadiSolver(pybamm.DaeSolver):
         self.extra_options = extra_options
         self.name = "CasADi solver ({}) with '{}' mode".format(method, mode)
 
-    def solve(self, model, t_eval):
+    def solve(self, model, t_eval, inputs=None):
         """
         Execute the solver setup and calculate the solution of the model at
         specified times.
@@ -80,6 +80,8 @@ class CasadiSolver(pybamm.DaeSolver):
             initial_conditions
         t_eval : numeric type
             The times at which to compute the solution
+        inputs : dict, optional
+            Any input parameters to pass to the model when solving
 
         Raises
         ------
@@ -99,7 +101,8 @@ class CasadiSolver(pybamm.DaeSolver):
         elif self.mode == "safe":
             # Step-and-check
             timer = pybamm.Timer()
-            self.set_up_casadi(model)
+            inputs = inputs or {}
+            self.set_up_casadi(model, inputs)
             set_up_time = timer.time()
             init_event_signs = np.sign(
                 np.concatenate([event(0, self.y0) for event in self.event_funs])
@@ -175,7 +178,7 @@ class CasadiSolver(pybamm.DaeSolver):
             )
             return solution
 
-    def compute_solution(self, model, t_eval):
+    def compute_solution(self, model, t_eval, inputs=None):
         """Calculate the solution of the model at specified times. In this class, we
         overwrite the behaviour of :class:`pybamm.DaeSolver`, since CasADi requires
         slightly different syntax.
@@ -187,6 +190,8 @@ class CasadiSolver(pybamm.DaeSolver):
             initial_conditions
         t_eval : numeric type
             The times at which to compute the solution
+        inputs : dict, optional
+            Any input parameters to pass to the model when solving
 
         """
         timer = pybamm.Timer()
@@ -198,6 +203,7 @@ class CasadiSolver(pybamm.DaeSolver):
             self.casadi_algebraic,
             self.y0,
             t_eval,
+            inputs,
             mass_matrix=model.mass_matrix.entries,
         )
         solve_time = timer.time() - solve_start_time
@@ -207,7 +213,9 @@ class CasadiSolver(pybamm.DaeSolver):
 
         return solution, solve_time, termination
 
-    def integrate_casadi(self, rhs, algebraic, y0, t_eval, mass_matrix=None):
+    def integrate_casadi(
+        self, rhs, algebraic, y0, t_eval, inputs=None, mass_matrix=None
+    ):
         """
         Solve a DAE model defined by residuals with initial conditions y0.
 
@@ -220,11 +228,14 @@ class CasadiSolver(pybamm.DaeSolver):
             The initial conditions
         t_eval : numeric type
             The times at which to compute the solution
+        inputs : dict, optional
+            Any input parameters to pass to the model when solving
         mass_matrix : array_like, optional
             The (sparse) mass matrix for the chosen spatial method. This is only passed
             to check that the mass matrix is diagonal with 1s for the odes and 0s for
             the algebraic equations, as CasADi does not allow to pass mass matrices.
         """
+        inputs = inputs or {}
         options = {
             "grid": t_eval,
             "reltol": self.rtol,
@@ -238,19 +249,15 @@ class CasadiSolver(pybamm.DaeSolver):
 
         # set up and solve
         t = casadi.MX.sym("t")
-        y_diff = casadi.MX.sym("y_diff", rhs(0, y0).shape[0])
+        u = casadi.vertcat(*[x for x in inputs.values()])
+        y_diff = casadi.MX.sym("y_diff", rhs(0, y0, u).shape[0])
+        problem = {"t": t, "x": y_diff}
         if algebraic is None:
-            problem = {"t": t, "x": y_diff, "ode": rhs(t, y_diff)}
+            problem.update({"ode": rhs(t, y_diff, u)})
         else:
-            y_alg = casadi.MX.sym("y_alg", algebraic(0, y0).shape[0])
+            y_alg = casadi.MX.sym("y_alg", algebraic(0, y0, u).shape[0])
             y = casadi.vertcat(y_diff, y_alg)
-            problem = {
-                "t": t,
-                "x": y_diff,
-                "z": y_alg,
-                "ode": rhs(t, y),
-                "alg": algebraic(t, y),
-            }
+            problem.update({"z": y_alg, "ode": rhs(t, y, u), "alg": algebraic(t, y, u)})
         integrator = casadi.integrator("F", self.method, problem, options)
         try:
             # Try solving